feat(leave-policy): add preview endpoint and support prorated entitlements
- Add PreviewLeaveAllocation handler to calculate prorated leave allocations without database access
- Support optional intern_entitlement parameter in policy preview calculations
- Change entitlement type from int to float64 across repositories and services to support prorated values
- Update leaveBalance service to prorate allocations based on policy creation date for new policies
- Enhance SyncLeaveBalances to preserve original joining dates during policy updates
- Add month parameter to preview endpoint for flexible allocation calculations
- Update PolicyFormDialog and policy step components to integrate preview functionality
- Enhance leaveService with preview method to fetch prorated allocation data

+++ backend/internal/repositories/leaveBalance.go
@@ -13,7 +13,7 @@ type leaveBalanceRepository struct {
 }
 
 type LeaveBalanceRepository interface {
-	Create(tx *sqlx.Tx, employeeID uuid.UUID, leaveTypeID int, entitlement int) error
+	Create(tx *sqlx.Tx, employeeID uuid.UUID, leaveTypeID int, entitlement float64) error
 	GetLeaveBalance(tx *sqlx.Tx, employeeID uuid.UUID, leaveTypeID int) (*models.LeaveBalanceForAdjustment, error)
 	UpdateLeaveBalance(tx *sqlx.Tx, balance *models.LeaveBalanceForAdjustment) error
 	UpdateAdjertment(tx *sqlx.Tx, balanceID uuid.UUID, newAdjusted, newClosing float64) error
@@ -25,7 +25,7 @@ func NewLeaveBalanceRepository(db *sqlx.DB) LeaveBalanceRepository {
 		DB: db,
 	}
 }
-func (r *leaveBalanceRepository) Create(tx *sqlx.Tx, employeeID uuid.UUID, leaveTypeID int, entitlement int) error {
+func (r *leaveBalanceRepository) Create(tx *sqlx.Tx, employeeID uuid.UUID, leaveTypeID int, entitlement float64) error {
 	_, err := tx.Exec(`
 		INSERT INTO Tbl_Leave_balance 
 			(employee_id, leave_type_id, year, opening, accrued, used, adjusted, closing)
